% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/allClasses.R
\name{glmdisc-class}
\alias{glmdisc-class}
\title{Class glmtree}
\description{
Class \code{glmtree} represents a logistic regression tree scheme associated with its optimal logistic regression models.
}
\section{Slots}{

\describe{
\item{\code{parameters}}{The parameters associated with the method.}

\item{\code{best.tree}}{The best discretization scheme found by the method given its parameters.}

\item{\code{performance}}{The performance obtained with the method given its parameters.}
}}

