% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialization.R
\name{initialization}
\alias{initialization}
\title{Random initialization of the segmentation}
\usage{
initialization(K, n)
}
\arguments{
\item{K}{The number of segments to start with (maximum number of segments there'll be in the end).}

\item{n}{The number of observations to draw.}
}
\value{
A vector of length n containing a factor with K levels (from 1 to K) drawn with replacement.
}
\description{
This function randomy initializes the segmentation.
}
\author{
Adrien Ehrhardt
}
\keyword{internal}
