% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_tweaked.R
\name{predict_tweaked}
\alias{predict_tweaked}
\title{Tweaked predict to take into account levels that are not in the training set but in the test set and remove these rows.}
\usage{
predict_tweaked(model, df, c_iter)
}
\arguments{
\item{model}{The logistic regression model to use to predict on a test set.}

\item{df}{The whole test set which class has to be predicted.}

\item{c_iter}{The segment considered (given by the tree).}
}
\value{
A dataframe of predictions (in rows: the observations, in cols: the class probabilities) given by the model given in input, eventually tweaked if some levels are unknown to it.
}
\description{
This function applies the logistic regression predict after carefully removing observations of categorical features' levels absent from the training set..
}
\author{
Adrien Ehrhardt
}
\keyword{internal}
